fix: fix linting error to return err value, move sendEmail to utils
